[FIX]: Moved the PropertiesForm.tsx file into Properties Folder to maintain proper folder structure
